+++ app/src/main/java/eu/jrie/lets_chess/Pieces/Queen.java
@@ -11,7 +11,7 @@ import java.util.ArrayList;
 import java.util.List;
 
 public class Queen extends Piece {
-    public Queen(Context c, int _color) {
+    public Queen(Context c, byte _color) {
         super(c, _color);
         type = Const.QUEEN;
         strokeColor = ContextCompat.getColor(context, R.color.colorQueen);
@@ -19,7 +19,7 @@ public class Queen extends Piece {
         else image = context.getResources().getDrawable(R.drawable.black_queen);
     }
 
-    public Queen(Context c, int _color, Position initPos) {
+    public Queen(Context c, byte _color, Position initPos) {
         super(c, _color, initPos);
         type = Const.QUEEN;
         strokeColor = ContextCompat.getColor(context, R.color.colorQueen);
